Rename Reap() to Clear() for clarity
Technically, "reap" or "reaping" has a particular definition associated
with it on UNIX systems, in particular, it refers to the act of calling
Wait() on a process and retrieving its exit information after its death,
culling it from the process table. Because this is a process management
framework, using Reap() as our name choice is kind of misleading to what
it actually does, which is _NOT_ actually reaping, but just clearing
dead processes in Jasper memory. Clear() is more honest about this.