Fix checkProcCapacity logic
The logic previously expected Clear() to throw an error _AND_ for us to
not have gained any new, needed space to throw an error and fail to
allocate another process. The correct logic here is to combine the two
with an _OR_ because either is a sufficient condition independently to
fail to allocate a process.