MAKE-528: add methods to clean up/clear terminated processes, also remove some redundant code (#20)
This commit adds functionality for a method on `Manager`s called `Clear()`, that will clean up already-completed/dead processes. It also adds an implementation for a `Manager` that has a limit on the number of alive+dead processes it stores in its memory and applies a `Clear()` call when needed for getting extra space.

Along with this, some no-longer useful code from a time before `basicProcess` was removed has been removed.