Propagate context logic from REST to JRPC
The JRPC code was not being as careful as the REST code implementation
for createProcess. This has been fixed by propagating the logic from
REST over to JRPC.

This may be an indication of a code-smell, since it seems a little too
easy to have the REST and JRPC implementations diverge somehow, but it
is also hard for me to see what we could do to fix this, really. It may
just be a price we must pay.