Add require.NoError() after manager.Get() calls
This causes seg faults in the tests if the returned process is nil and
there is an error. Technically, this doesn't actually "fix" the error,
but it at least will avoid seg faults that destroy a large part of the
test and give a more helpful error output in the logs. This is not
related to MAKE-569, and is actually related to MAKE-572.