Further improve basicProcess
Clean up the "reactor()" code by using better names, and remove the
unnecessary sig channel for ensuring that Wait() is called. Overall,
this finalizes the merging of our own idea of splitting the process
lifecycle into a set of phases it undergoes, maintaining an info that is
consistent at all times and stored in the process itself, and defining
Complete() in terms of Running().