Fix flaky test: SignalErrorsWithInvalidSyscall
Since we were using a very fast command (trueCreateOpts()) it was always
possible that the `true` command would terminate before we even got
around to sending the signal. In the case that it did, the call to
cmd.Process.Signal() would actually not return an error since the
process was already dead. This ensures that we don't run into this race
by using a much longer command (sleep 10).