MAKE-537: Add some missing tests for basicProcess in JRPC & integration_test.go (#30)
This commit adds some code that ensures that basicProcess AND blockingProcess is tested elsewhere in the code, such as the JRPC tests and integration tests. It also changes slightly the expected behavior of Wait() to now return the error returned from Go's stdlib Wait(), instead of returning a custom, more-vague error upon unsuccessful executions of processes.