EVG-19927: Code review cleanup and another data race.
This change updates the docker executor to make the wait group for
io accessible on the executor itself. This allows executor_test to
wait until io is done and avoid data races.

Additionally, this fixes a bug in process_blocking where a write
was done under a read lock, and updates a test in manager_test
where checking not empty accessed the underlying object and caused
a data race.