EVG-19927: Upgrade go and Evergreen distros. (#336)
* EVG-19927: Upgrade go and Evergreen distros.

This upgrades to go 1.20, and makes sure we use newer versions
of our Evergreen distros.

*  Fix linter errors.

This required refactoring some of oom_tracker.go into a helper
file that builds only on darwin/Linux.

* Don't use mock process as "invalid input".

We were only serializing mock process to be "invalid input" for
the cli manager test. We should just use an arbitrary map instead.

*  Switch executor_test to use SafeBuffer.

This is the simplest way to avoid data races.